requirements.txt
fastapi
uvicorn
tortoise-orm[asyncpg]
aerich
google-api-python-client
google-auth
google-auth-oauthlib
google-genai
chromadb
python-dotenv
pydantic
httpx
email-validator